// frontend/src/app/core/api/model/exam-schedule-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExamScheduleRequest { 
    assessmentId: number;
    examDate: string;
    examTimeSlotId: number;
    roomId?: number;
    capacityUsed?: number;
}